# Convert `z` in `CubicSpline` for performance