--- a/src/interpolation_methods.jl
+++ b/src/interpolation_methods.jl
@@ -349,7 +349,7 @@ end
 
 # Cubic Hermite Spline
 function _interpolate(
-        A::CubicHermiteSpline{<:AbstractVector{<:Number}}, t::Number, iguess)
+        A::CubicHermiteSpline{<:AbstractVector}, t::Number, iguess)
     idx = get_idx(A, t, iguess)
     Δt₀ = t - A.t[idx]
     Δt₁ = t - A.t[idx + 1]