--- a/test/interpolation_tests.jl
+++ b/test/interpolation_tests.jl
@@ -933,6 +933,7 @@ end
     push!(u, 1.0)
     @test_throws AssertionError CubicHermiteSpline(du, u, t)
 
+    deleteat!(u, lastindex(u))
     @testset "Vector of Vectors case" begin
         u2 = [[u[i], u[i] + 1] for i in eachindex(u)]
         du2 = [[du[i], du[i]] for i in eachindex(du)]
@@ -942,6 +943,7 @@ end
     @testset "Vector of Matrices case" begin
         u3 = [[u[i] u[i] + 1] for i in eachindex(u)]
         du3 = [[du[i] du[i]] for i in eachindex(du)]
+        @test length(u3) == length(du3)
         A3 = CubicHermiteSpline(du3, u3, t)
         @test u3 ≈ A3.(t)
     end
@@ -981,6 +983,7 @@ end
     @test @inferred(output_dim(A)) == 0
     @test @inferred(output_size(A)) == ()
 
+    deleteat!(u, lastindex(u))
     @testset "Vector of Vectors case" begin
         u2 = [[u[i], u[i] + 1] for i in eachindex(u)]
         du2 = [[du[i], du[i]] for i in eachindex(du)]