--- a/test/interpolation_tests.jl
+++ b/test/interpolation_tests.jl
@@ -88,8 +88,7 @@ end
         # Test allocation-free interpolation with StaticArrays
         u_s = [convert(SVector{length(y), eltype(u_)}, i) for i in u]
         @test @inferred(LinearInterpolation(
-            u_s, t; extrapolation = ExtrapolationType.Extension)) isa LinearInterpolation broken=VERSION <
-                                                                                               v"1.11" &&
+            u_s, t; extrapolation = ExtrapolationType.Extension)) isa LinearInterpolation &&
                                                                                                t isa
                                                                                                AbstractRange
         A_s = LinearInterpolation(u_s, t; extrapolation = ExtrapolationType.Extension)