## Checklist

- [ ] Appropriate tests were added
- [ ] Any code changes were done in a way that does not break public API
- [ ] All documentation related to code changes were updated
- [ ] The new code follows the
  [contributor guidelines](https://github.com/SciML/.github/blob/master/CONTRIBUTING.md), in particular the [SciML Style Guide](https://github.com/SciML/SciMLStyle) and
  [COLPRAC](https://github.com/SciML/COLPRAC).
- [ ] Any new documentation only uses public API
  
## Additional context

This hopes to address the performance issues of multi-valued interpolation discussed in #431 by using dependent variable arrays with elements of `StaticArrays.SArray` to remove allocations and improve performance. On `master` this still allocates when evaluating the interpolants, hurting performance.

For the example of `CubicSpline`, I isolated the allocations to come from the `z` array. By converting the elements of `z` to match `eltype(u)`, we are able to remove allocations when evaluating the interpolant and greatly improve the performance. Example usage and performance comparison is given below. 

All tests pass for me locally on Julia 1.11.6. If this seems like a reasonable solution, I can look into modifying the other relevant methods as well.

```julia
using StaticArrays: SVector, SArray
using DataInterpolations: CubicSpline
using BenchmarkTools: @benchmark
using  Test: @test

x = 1:1_000
a,b,c = [rand(length(x)) for i in 1:3]
y = [SVector{3, Float64}(a[i], b[i], c[i]) for i in eachindex(a,b,c)]

cubic = CubicSpline(y, x)
@benchmark $cubic($2.5)
# On master, ↑ is 133.750 ns with Memory estimate: 640 bytes, allocs estimate: 16.
# On PR: 21.717 ns, 0 allocations

y2 = [SArray{Tuple{1,2,3}, Float64, 3, 6}(rand(2,3)) for i in eachindex(a,b,c)]
cubic2 = CubicSpline(y2, x)
@benchmark $cubic2($2.5)
# On master, ↑ is 248.704 ns with Memory estimate: 1.00 KiB, allocs estimate: 16.
# On PR: 25.965 ns ns, 0 allocations

# Vector of vectors
y3 = [[a[i], b[i], c[i]] for i in eachindex(a,b,c)]
cubic3 = CubicSpline(y3, x)
@benchmark $cubic3($2.5)
# On master, ↑ is 243.811 ns with Memory estimate: 1.17 KiB, allocs estimate: 30.
# On PR: 229.563 ns with Memory estimate: 1.17 KiB, allocs estimate: 30.

# For Matrix y
y4 = [a b c]'
x4 = 1:size(y4, 2)
cubic4 = CubicSpline(y4, x4)
@benchmark $cubic4($2.5)
# On master, ↑ is 307.918 ns with Memory estimate: 1.64 KiB, allocs estimate: 42.
@test cubic(2.5) == cubic4(2.5) # Gives same result
```