@ChrisRackauckas 2025-08-10
It looks reasonable enough to me. Though what does it do to the arrays of arrays case? allocate the views?