@cgarling 2025-08-10
I think so, results for `y::Vector{Vector}` (named `y3`) added above