@cgarling 2025-08-11
With this PR, the following methods work with `u::AbstractVector{<:AbstractVector}`, and seem to have good performance and no allocations when `u::AbstractVector{<:StaticArrays.SVector}`. 
 - CubicSpline
 - QuadraticInterpolation
 - SmoothedConstantInterpolation
 - ConstantInterpolation
 - QuadraticSpline
 - LinearInterpolation
 - CubicHermiteSpline
 - QuinticHermiteSpline

Not all of these methods have tests confirming this behavior. I had to relax the type signatures of the `_interpolate` methods for `CubicHermiteSpline` and `QuinticHermiteSpline` to enable this behavior, so I added tests confirming that the alterations result in successful operation. I could look at adding similar extra tests to the other interpolation types above, if you'd like to have this behavior covered by the tests, but that should probably be under a different PR.

The remaining methods that don't seem to work with  `u::AbstractVector{<:AbstractVector}` are
- PCHIPInterpolation
- AkimaInterpolation
- LagrangeInterpolation
- BSplineInterpolation
- BSplineApprox

I am interested in trying to get PCHIPInterpolation and AkimaInterpolation working, so I may look into that in another PR. For now I think this PR is ready for review.